+++ src/coreComponents/physicsSolvers/fluidFlow/CompositionalMultiphaseFVM.hpp
@@ -26,6 +26,26 @@ namespace geos
 {
 
 /**
+ * @brief Options for density treatment in gravity
+ */
+enum class GravityDensityScheme : integer
+{
+  ArithmeticAverage, ///< average phase density is computed using simple arithmetic average:
+                     ///  rho_ave = 0.5 * (rho_i + rho_j)
+  PhasePresence,     ///< average phase density is computed using checking for phase presence:
+                     ///  rho_ave = 0.5 * (rho_i + rho_j) if phase is present in both cells i and j
+                     ///          = rho_i if phase is present in only cell i
+                     ///          = rho_j if phase is present in only cell j
+};
+
+/**
+ * @brief Strings for options for density treatment in gravity
+ */
+ENUM_STRINGS( GravityDensityScheme,
+              "ArithmeticAverage",
+              "PhasePresence" );
+
+/**
  * @class CompositionalMultiphaseFVM
  *
  * A compositional multiphase solver
@@ -84,6 +104,11 @@ public:
    * These functions provide the primary interface that is required for derived classes
    */
   /**@{*/
+
+  virtual void
+  registerDataOnMesh( Group & MeshBodies ) override;
+
+  virtual void registerDataForCFL( Group & meshBodies ) override;
 
   virtual void
   setupDofs( DomainPartition const & domain,
@@ -150,6 +175,20 @@ public:
                   CRSMatrixView< real64, globalIndex const > const & localMatrix,
                   arrayView1d< real64 > const & localRhs ) const override;
 
+  virtual void computeCFLNumbers( DomainPartition & domain,
+                                  real64 const & dt,
+                                  real64 & maxPhaseCFL,
+                                  real64 & maxCompCFL ) override;
+
+  /**
+   * @brief function to set the next time step size
+   * @param[in] currentDt the current time step size
+   * @param[in] domain the domain object
+   * @return the prescribed time step size
+   */
+  real64 setNextDt( real64 const & currentDt,
+                    DomainPartition & domain ) override;
+
   struct viewKeyStruct : CompositionalMultiphaseBase::viewKeyStruct
   {
     // DBC parameters
@@ -162,7 +201,8 @@ public:
     static constexpr char const * contMultiplierDBCString()       { return "contMultiplierDBC"; }
 
     // nonlinear solver parameters
-    static constexpr char const * scalingTypeString()               { return "scalingType"; }
+    static constexpr char const * scalingTypeString()             { return "scalingType"; }
+    static constexpr char const * gravityDensitySchemeString()    { return "gravityDensityScheme"; }
   };
 
   /**
@@ -191,8 +231,10 @@ protected:
 
   virtual void postInputInitialization() override;
 
-  virtual void
-  initializePreSubGroups() override;
+  virtual void initializePreSubGroups() override;
+
+  real64 setNextDtBasedOnCFL( real64 const & currentDt,
+                              DomainPartition & domain );
 
   struct DBCParameters
   {
@@ -214,6 +256,12 @@ protected:
 
   /// Solution scaling type
   ScalingType m_scalingType;
+
+  /// scheme for density treatment in gravity
+  GravityDensityScheme m_gravityDensityScheme;
+
+  /// the targeted CFL for timestep
+  real64 m_targetFlowCFL;
 
 private:
