+++ src/coreComponents/physicsSolvers/fluidFlow/CompositionalMultiphaseStatistics.cpp
@@ -142,22 +142,13 @@ void CompositionalMultiphaseStatistics::registerDataOnMesh( Group & meshBodies )
         }
       }
     }
-
-    // if we have to compute CFL numbers later, we need to register additional variables
-    if( m_computeCFLNumbers )
-    {
-      elemManager.forElementSubRegions( regionNames, [&]( localIndex const,
-                                                          ElementSubRegionBase & subRegion )
-      {
-        subRegion.registerField< fields::flow::phaseOutflux >( getName() ).
-          reference().resizeDimension< 1 >( numPhases );
-        subRegion.registerField< fields::flow::componentOutflux >( getName() ).
-          reference().resizeDimension< 1 >( numComps );
-        subRegion.registerField< fields::flow::phaseCFLNumber >( getName() );
-        subRegion.registerField< fields::flow::componentCFLNumber >( getName() );
-      } );
-    }
   } );
+
+  // if we have to compute CFL numbers later, we need to register additional variables
+  if( m_computeCFLNumbers )
+  {
+    m_solver->registerDataForCFL( meshBodies );
+  }
 }
 
 bool CompositionalMultiphaseStatistics::execute( real64 const time_n,
